put q_grid function in tools.py
