+++ src/ess/amor/__init__.py
@@ -1,2 +1,3 @@
 # flake8: noqa: F401
-from ess.amor.amor_data import AmorData, AmorReference, Normalisation, q_grid
+from ess.amor.amor_data import AmorData, AmorReference, Normalisation
+from ess.amor.tools import q_grid
